Made pagos faster, started stock
The embarques file is no longer openned every time the ingresar button gets pressed, instead, it checks if the embarque file has been updated since the up has started. It should check if it has been modified since the program last openned it (for more efficient) but, since this is an edge case, tough luck.